enrichment-worker/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
beautifulsoup4>=4.13.0
lxml>=5.0.0
phonenumbers>=8.13.0
email-validator>=2.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
